ci/cd round 2: make the whole matrix actually pass
Reproduced and fixed every failing leg locally (the runner failures were
all real, x86-only blind spots):

- hw_deltanet NEON: kiv was initialized with dec instead of ki
  (copy-paste vs the AVX512 path) — dn_rows failed under qemu on every
  ARM tier. All 4 tiers + armv7 (A7/A15) now ALL PASS under qemu.
- Makefile: LTO archives built with $(CC)-ar (plain binutils ar can
  drop LTO bytecode -> undefined refs on the runner's gcc-13; this was
  the 'Build engines' failure). ENG_OBJS was defined after first use —
  moty-plugins linked without the engine objects.
- Windows/MinGW leg compiles now: rss_gb guarded behind
  MOTY_HAVE_RUSAGE (reporting-only fallback), MADV_HUGEPAGE calls
  behind MOTY_MADVISE, local IN renamed (WinSDK macro collision) in
  gguf.h/qwen.c/qwen_train.h, strndup shim in gguf.h (msvcrt),
  mallopt(M_MMAP_THRESHOLD) guarded to glibc.
- ci.yml plugins job: uses 'make moty-plugins' (same flags as the
  tree) instead of a hand-rolled gcc line.
- release.yml: armv7 package test counts files in the dir (no moty on
  that triplet by contract).

Gate locally: 118/118 portable+native, all four release targets build
clean, LFM2 56.2 / Qwen 20.7 tok/s VERDE.

+++ c/nn/alloc.c
@@ -7,7 +7,7 @@ void *moty_grow(void **p, int64_t *cap, int64_t need, size_t esz, const char *wh
     if (need <= *cap) return *p;
     *cap = need;
     *p = realloc(*p, (size_t)need * esz);
-    if (!*p) { fprintf(stderr, "OOM %s (%lld x %zu byte)\n", what, (long long)need, esz); exit(1); }
+    if (!*p) { fprintf(stderr, "OOM %s (%lld x %lu byte)\n", what, (long long)need, (unsigned long)esz); exit(1); }
     return *p;
 }
 void *moty_balloc(int64_t n, const char *what) {
